split createSliceWithImmer test
