+++ tests/01_basic.spec.tsx
@@ -1,11 +1,9 @@
 import { expect, test } from 'vitest';
 import { create } from 'zustand';
 import { createSlice, withSlices } from 'zustand-slices';
-import { createSliceWithImmer } from 'zustand-slices/immer';
 
 test('should export functions', () => {
   expect(createSlice).toBeDefined();
-  expect(createSliceWithImmer).toBeDefined();
   expect(withSlices).toBeDefined();
 });
 
@@ -19,29 +17,6 @@ test('createSlice', () => {
   });
   // returns the input
   expect(createSlice(slice)).toBe(slice);
-});
-
-test('createSliceWithImmer', () => {
-  const immerSlice = createSliceWithImmer({
-    name: 'counter',
-    value: {
-      count: 0,
-    },
-    actions: {
-      increment: () => (prev) => {
-        prev.count += 1;
-      },
-    },
-  });
-  const result = createSliceWithImmer(immerSlice);
-
-  // should not be equal as createSliceWithImmer should wrap actions in `produce`
-  expect(result.actions.increment).not.toBe(immerSlice.actions.increment);
-
-  expect(result.name).toEqual(immerSlice.name);
-  expect(result.value).toEqual(immerSlice.value);
-
-  expect(typeof result.actions.increment).toBe('function');
 });
 
 test('withSlices', () => {
